refactored out /object/ routes from server.js
