+++ server.js
@@ -86,6 +86,7 @@ const globalVariables = {
     developer: true,
     // Send more debug messages to console
     debug: false,
+    isMobile: isMobile,
     // Prohibit saving to file system if we're on mobile or just running tests
     saveToDisk: !isMobile && process.env.NODE_ENV !== 'test',
     // Create an object for attaching frames to the world
@@ -1508,6 +1509,7 @@ function objectWebServer() {
         }
     });
 
+    // TODO: is the developer flag ever not true anymore? is it still useful to have?
     if (globalVariables.developer === true) {
         webServer.use('/libraries', express.static(__dirname + '/libraries/webInterface/'));
         webServer.use('/libraries/monaco-editor/', express.static(__dirname + '/node_modules/monaco-editor/'));
@@ -1604,297 +1606,51 @@ function objectWebServer() {
         });
 
     }
+    
+    // Express router routes
+    const objectRouter = require('./routers/object');
+    objectRouter.setup(objects, knownObjects, socketArray, globalVariables, engine, hardwareAPI, __dirname, objectsPath, identityFolderName, Jimp, socketUpdater, git, nodeTypeModules);
+    webServer.use('/object', objectRouter.router);
 
-
-    /// logic node handling
-
-
+    // Controller imports
+    const blockController = require('./controllers/block.js');
+    const blockLinkController = require('./controllers/blockLink.js');
+    const logicNodeController = require('./controllers/logicNode.js');
+    
+    // logic node handling
     /**
      * Logic Links
      **/
-
-    // support for frames
-    webServer.delete('/object/*/frame/*/node/*/link/*/editor/*/deleteBlockLink/', function (req, res) {
-        res.send(deleteLogicLink(req.params[0], req.params[1], req.params[2], req.params[3], req.params[4]));
-    });
-
-
     // delete a logic link. *1 is the object *2 is the logic *3 is the link id
-    // ****************************************************************************************************************
     webServer.delete('/logic/*/*/link/*/lastEditor/*/', function (req, res) {
-        res.send(deleteLogicLink(req.params[0], req.params[0], req.params[1], req.params[2], req.params[3]));
+        res.send(blockLinkController.deleteLogicLink(objects, globalVariables, objectsPath, req.params[0], req.params[0], req.params[1], req.params[2], req.params[3]));
     });
-
-    function deleteLogicLink(objectID, frameID, nodeID, linkID, lastEditor) {
-
-        var updateStatus = 'nothing happened';
-
-        var foundNode = getNode(objectID, frameID, nodeID);
-        if (foundNode) {
-            delete foundNode.links[linkID];
-
-            utilities.actionSender({
-                reloadNode: {object: objectID, frame: frameID, node: nodeID},
-                lastEditor: lastEditor
-            });
-            utilities.writeObjectToFile(objects, objectID, objectsPath, globalVariables.saveToDisk);
-
-            console.log('deleted link: ' + linkID);
-            updateStatus = 'deleted: ' + linkID + ' in logic ' + nodeID + ' in frame: ' + frameID + ' from object: ' + objectID;
-        }
-
-        return updateStatus;
-
-    }
-
-
+    
     // adding a new logic link to an object. *1 is the object *2 is the logic *3 is the link id
-    //
-    // ****************************************************************************************************************
-
     webServer.post('/logic/*/*/link/*/', function (req, res) {
-        res.send(addLogicLink(req.params[0], req.params[0], req.params[1], req.params[2], req.body));
+        res.send(blockLinkController.addLogicLink(objects, globalVariables, objectsPath, req.params[0], req.params[0], req.params[1], req.params[2], req.body));
     });
-    // support for frames
-    webServer.post('/object/*/frame/*/node/*/link/*/addBlockLink/', function (req, res) {
-        res.send(addLogicLink(req.params[0], req.params[1], req.params[2], req.params[3], req.body));
-    });
-
-    /**
-     * Adds a new link with the provided linkID to the specified node.
-     * Doesn't add it if it detects an infinite loop.
-     * @param {string} objectID
-     * @param {string} frameID
-     * @param {string} nodeID
-     * @param {string} linkID
-     * @param {Link} body
-     * @return {string}
-     */
-    function addLogicLink(objectID, frameID, nodeID, linkID, body) {
-
-        var updateStatus = 'nothing happened';
-
-        var foundNode = getNode(objectID, frameID, nodeID);
-        if (foundNode) {
-
-            foundNode.links[linkID] = body;
-            var thisLink = foundNode.links[linkID];
-
-            thisLink.loop = false;
-            // todo the first link in a chain should carry a UUID that propagates through the entire chain each time a change is done to the chain.
-            // todo endless loops should be checked by the time of creation of a new loop and not in the Engine
-            if (thisLink.nodeA === thisLink.nodeB && thisLink.logicA === thisLink.logicB) {
-                thisLink.loop = true;
-            }
-
-            if (!thisLink.loop) {
-                // call an action that asks all devices to reload their links, once the links are changed.
-                utilities.actionSender({
-                    reloadNode: {object: objectID, frame: frameID, node: nodeID},
-                    lastEditor: body.lastEditor
-                });
-                // check if there are new connections associated with the new link.
-                // write the object state to the permanent storage.
-                utilities.writeObjectToFile(objects, objectID, objectsPath, globalVariables.saveToDisk);
-
-                console.log('added link: ' + linkID);
-                updateStatus = 'added';
-            } else {
-                updateStatus = 'found endless Loop';
-            }
-        }
-
-        return updateStatus;
-    }
 
     /**
      * Logic Blocks
      **/
-
     // adding a new block to an object. *1 is the object *2 is the logic *3 is the link id
-    // ****************************************************************************************************************
-    webServer.post('/object/:objectID/frame/:frameID/node/:nodeID/block/:blockID/addBlock/', function (req, res) {
-        res.send(addNewBlock(req.params.objectID, req.params.frameID, req.params.nodeID, req.params.blockID, req.body));
-    });
-
     webServer.post('/logic/*/*/block/*/', function (req, res) {
-        res.send(addNewBlock(req.params[0], req.params[0], req.params[1], req.params[2], req.body));
+        res.send(blockController.addNewBlock(objects, globalVariables, objectsPath, req.params[0], req.params[0], req.params[1], req.params[2], req.body));
     });
-
-    /**
-     * Adds a new block with the provided blockID to the specified node.
-     * @param {string} objectID
-     * @param {string} frameID
-     * @param {string} nodeID
-     * @param {string} blockID
-     * @param {Block} body
-     * @return {string}
-     */
-    function addNewBlock(objectID, frameID, nodeID, blockID, body) {
-
-        var updateStatus = 'nothing happened';
-
-        var foundNode = getNode(objectID, frameID, nodeID);
-        if (foundNode) {
-
-            var thisBlocks = foundNode.blocks;
-            thisBlocks[blockID] = new Block();
-
-            // todo activate when system is working to increase security
-            /* var thisMessage = req.body;
-
-             var thisModule = {};
-
-             var breakPoint = false;
-
-             if (thisMessage.type in blockFolderList) {
-             thisModule = blockModules[thisMessage.type];
-
-             for (var thisKey in thisMessage.publicData) {
-             if (typeof thisMessage.publicData[thisKey] !== typeof thisModule.publicData[thisKey]) {
-             breakPoint = true;
-             }
-             }
-
-             for (var thisKey in thisMessage.privateData) {
-             if (typeof thisMessage.privateData[thisKey] !== typeof thisModule.privateData[thisKey]) {
-             breakPoint = true;
-             }
-             }
-             }
-             else {
-             breakPoint = true;
-             }
-
-             if (!breakPoint)*/
-
-            thisBlocks[blockID] = body;
-
-            // todo this can be removed once the system runs smoothly
-            if (typeof thisBlocks[blockID].type === 'undefined') {
-                thisBlocks[blockID].type = thisBlocks[blockID].name;
-            }
-
-            // call an action that asks all devices to reload their links, once the links are changed.
-            utilities.actionSender({
-                reloadNode: {object: objectID, frame: frameID, node: nodeID},
-                lastEditor: body.lastEditor
-            });
-            utilities.writeObjectToFile(objects, objectID, objectsPath, globalVariables.saveToDisk);
-
-            console.log('added block: ' + blockID);
-            updateStatus = 'added';
-        }
-
-        return updateStatus;
-    }
 
     // delete a block from the logic. *1 is the object *2 is the logic *3 is the link id
-    // ****************************************************************************************************************
-    webServer.delete('/object/:objectID/frame/:frameID/node/:nodeID/block/:blockID/editor/:lastEditor/deleteBlock/', function (req, res) {
-        res.send(deleteBlock(req.params.objectID, req.params.frameID, req.params.nodeID, req.params.blockID, req.params.lastEditor));
-    });
     // webServer.delete('/logic/*/*/*/block/*/lastEditor/*/', function (req, res) {
-    //     res.send(deleteBlock(req.params[0], req.params[1], req.params[2], req.params[3], req.params[4]));
+    //     res.send(blockController.deleteBlock(objects, globalVariables, objectsPath, req.params[0], req.params[1], req.params[2], req.params[3], req.params[4]));
     // });
 
-    /**
-     * Deletes a block with the provided blockID from the the specified node.
-     * Also deletes any links connected to that block.
-     * @param {string} objectID
-     * @param {string} frameID
-     * @param {string} nodeID
-     * @param {string} blockID
-     * @param {string} lastEditor
-     * @return {string}
-     */
-    function deleteBlock(objectID, frameID, nodeID, blockID, lastEditor) {
-
-        var updateStatus = 'nothing happened';
-
-        var foundNode = getNode(objectID, frameID, nodeID);
-
-        if (foundNode) {
-
-            delete foundNode.blocks[blockID];
-            console.log('deleted block: ' + blockID);
-
-            var thisLinks = foundNode.links;
-            // Make sure that no links are connected to deleted blocks
-            for (var linkCheckerKey in thisLinks) {
-                if (!thisLinks.hasOwnProperty(linkCheckerKey)) continue;
-                if (thisLinks[linkCheckerKey].nodeA === blockID || thisLinks[linkCheckerKey].nodeB === blockID) { // TODO: do we need to check blockLinks?
-                    delete foundNode.links[linkCheckerKey];
-                }
-            }
-
-            utilities.actionSender({
-                reloadNode: {object: objectID, frame: nodeID, node: nodeID},
-                lastEditor: lastEditor
-            });
-            utilities.writeObjectToFile(objects, objectID, objectsPath, globalVariables.saveToDisk);
-            updateStatus = 'deleted: ' + blockID + ' in blocks for object: ' + objectID;
-        }
-
-        return updateStatus;
-    }
-
     webServer.post('/logic/*/*/blockPosition/*/', function (req, res) {
-        res.send(postBlockPosition(req.params[0], req.params[0], req.params[1], req.params[2], req.body));
+        res.send(blockController.postBlockPosition(objects, globalVariables, objectsPath, req.params[0], req.params[0], req.params[1], req.params[2], req.body));
     });
-
-    webServer.post('/object/*/frame/*/node/*/block/*/blockPosition/', function (req, res) {
-        res.send(postBlockPosition(req.params[0], req.params[1], req.params[2], req.params[3], req.body));
-    });
-
-    /**
-     * Sets a new grid position for the specified block
-     * @param {string} objectID
-     * @param {string} frameID
-     * @param {string} nodeID
-     * @param {string} blockID
-     * @param {{x: number, y: number, lastEditor: string}} body
-     * @return {string}
-     */
-    function postBlockPosition(objectID, frameID, nodeID, blockID, body) {
-
-        var updateStatus = 'nothing happened';
-
-        console.log('changing Position for :' + objectID + ' : ' + nodeID + ' : ' + blockID);
-
-        var foundNode = getNode(objectID, frameID, nodeID);
-
-        if (foundNode) {
-            var foundBlock = foundNode.blocks[blockID];
-            if (foundBlock) {
-                // check that the numbers are valid numbers..
-                if (typeof body.x === 'number' && typeof body.y === 'number') {
-
-                    foundBlock.x = body.x;
-                    foundBlock.y = body.y;
-
-                    utilities.writeObjectToFile(objects, objectID, objectsPath, globalVariables.saveToDisk);
-                    utilities.actionSender({
-                        reloadNode: {object: objectID, frame: frameID, node: nodeID},
-                        lastEditor: body.lastEditor
-                    });
-                    updateStatus = 'ok';
-                }
-            }
-        }
-
-        return updateStatus;
-    }
 
     // receivePost blocks can be triggered with a post request. *1 is the object *2 is the logic *3 is the link id
-    // ****************************************************************************************************************
-    webServer.post('/object/:objectID/frame/:frameID/node/:nodeID/block/:blockID/triggerBlock/', function (req, res) {
-        res.send(triggerBlock(req.params.objectID, req.params.frameID, req.params.nodeID, req.params.blockID, req.body));
-    });
-
     // abbreviated POST syntax, searches over all objects and frames to find the block with that ID
     webServer.post('/triggerBlock/:blockID', function (req, res) {
-
         var foundBlock = false;
         forEachObject(function (object, objectKey) {
             forEachFrameInObject(object, function (frame, frameKey) {
@@ -1904,203 +1660,35 @@ function objectWebServer() {
                         // keep iterating until you find a block with that ID
                         if (block) {
                             foundBlock = true;
-                            res.status(200).json(triggerBlock(objectKey, frameKey, nodeKey, req.params.blockID, req.body)).end();
+                            res.status(200).json(blockController.triggerBlock(objects, engine, objectKey, frameKey, nodeKey, req.params.blockID, req.body)).end();
                         }
                     }
                 });
             });
         });
-
         if (!foundBlock) {
             res.status(404).json({success: false, error: 'no block with ID ' + req.params.blockID + ' exists'}).end();
         }
-
     });
-
-    function triggerBlock(objectID, frameID, nodeID, blockID, body) {
-        console.log('triggerBlock', objectID, frameID, nodeID, blockID, body);
-        var foundNode = getNode(objectID, frameID, nodeID);
-        if (foundNode) {
-            var block = foundNode.blocks[blockID];
-            console.log('block', block);
-            console.log('set block ' + block.type + ' (' + blockID + ') to ' + body.value);
-
-            block.data[0].value = body.value;
-            engine.blockTrigger(objectID, frameID, nodeID, blockID, 0, block);
-        }
-
-        return {success: true, error: null};
-    }
 
     /**
      * Logic Nodes
      **/
-
     // adding a new logic node block to an object. *1 is the object *2 is the logic *3 is the link id
-    // ****************************************************************************************************************
     webServer.post('/logic/*/*/node/', function (req, res) {
-        res.send(addLogicNode(req.params[0], req.params[0], req.params[1], req.body));
+        res.send(logicNodeController.addLogicNode(objects, globalVariables, objectsPath, req.params[0], req.params[0], req.params[1], req.body));
     });
-
-    webServer.post('/object/*/frame/*/node/*/addLogicNode/', function (req, res) {
-        res.send(addLogicNode(req.params[0], req.params[1], req.params[2], req.body));
-    });
-
-    /**
-     * Adds the Logic Node contained in the body to the specified frame.
-     * Creates some state (edge blocks) necessary for the server data processing that doesn't exist in the client.
-     * @param {string} objectID
-     * @param {string} frameID
-     * @param {string} nodeID
-     * @param {Node} body
-     * @return {string}
-     */
-    function addLogicNode(objectID, frameID, nodeID, body) {
-        var updateStatus = 'nothing happened';
-
-        var foundFrame = getFrame(objectID, frameID);
-        if (foundFrame) {
-
-            foundFrame.nodes[nodeID] = body;
-            var newNode = foundFrame.nodes[nodeID];
-
-            // edge blocks are used to transition data between node links going into the red/yellow/green/blue ports...
-            // ...and the corresponding blocks / block links within the crafting board
-            newNode.blocks['in0'] = new EdgeBlock();
-            newNode.blocks['in1'] = new EdgeBlock();
-            newNode.blocks['in2'] = new EdgeBlock();
-            newNode.blocks['in3'] = new EdgeBlock();
-
-            newNode.blocks['out0'] = new EdgeBlock();
-            newNode.blocks['out1'] = new EdgeBlock();
-            newNode.blocks['out2'] = new EdgeBlock();
-            newNode.blocks['out3'] = new EdgeBlock();
-
-            newNode.type = 'logic';
-
-            // call an action that asks all devices to reload their links, once the links are changed.
-            utilities.writeObjectToFile(objects, objectID, objectsPath, globalVariables.saveToDisk);
-            utilities.actionSender({
-                reloadNode: {object: objectID, frame: frameID, node: nodeID},
-                lastEditor: body.lastEditor
-            });
-
-            console.log('added logic node: ' + nodeID);
-            updateStatus = 'added';
-        }
-
-        return updateStatus;
-    }
-
+    
     // delete a logic node from the logic. *1 is the object *2 is the logic *3 is the link id
-    // ****************************************************************************************************************
     webServer.delete('/logic/*/*/node/lastEditor/*/', function (req, res) {
-        res.send(deleteLogicNode(req.params[0], req.params[0], req.params[1], req.params[2]));
+        res.send(logicNodeController.deleteLogicNode(objects, globalVariables, objectsPath, req.params[0], req.params[0], req.params[1], req.params[2]));
     });
-
-    webServer.delete('/object/*/frame/*/node/*/editor/*/deleteLogicNode', function (req, res) {
-        res.send(deleteLogicNode(req.params[0], req.params[1], req.params[2], req.params[3]));
-    });
-
-    /**
-     * Deletes the specified Logic Node.
-     * @param {string} objectID
-     * @param {string} frameID
-     * @param {string} nodeID
-     * @param {string} lastEditor
-     * @return {string}
-     */
-    function deleteLogicNode(objectID, frameID, nodeID, lastEditor) {
-
-        var updateStatus = 'nothing happened';
-
-        var foundFrame = getFrame(objectID, frameID);
-        if (foundFrame) {
-            delete foundFrame.nodes[nodeID];
-            console.log('deleted node: ' + nodeID);
-
-            //todo check all links as well in object
-            // Make sure that no links are connected to deleted objects
-            /*  for (var subCheckerKey in  objects[req.params[0]].links) {
-
-                  if (objects[req.params[0]].links[subCheckerKey].nodeA === req.params[1] && objects[req.params[0]].links[subCheckerKey].objectA === req.params[0]) {
-                      delete objects[req.params[0]].links[subCheckerKey];
-                  }
-                  if (objects[req.params[0]].links[subCheckerKey].nodeB === req.params[1] && objects[req.params[0]].links[subCheckerKey].objectB === req.params[0]) {
-                      delete objects[req.params[0]].links[subCheckerKey];
-                  }
-              }*/
-
-            utilities.writeObjectToFile(objects, objectID, objectsPath, globalVariables.saveToDisk);
-            utilities.actionSender({
-                reloadNode: {object: objectID, frame: frameID, node: nodeID},
-                lastEditor: lastEditor
-            });
-
-            updateStatus = 'deleted: ' + nodeID + ' in frame: ' + frameID + ' of object: ' + objectID;
-        }
-
-        return updateStatus;
-    }
 
     webServer.post('/logic/*/*/nodeSize/', function (req, res) {
-        changeNodeSize(req.params[0], req.params[0], req.params[1], req.body, function (statusCode, responseContents) {
+        logicNodeController.changeNodeSize(objects, globalVariables, objectsPath, req.params[0], req.params[0], req.params[1], req.body, function (statusCode, responseContents) {
             res.status(statusCode).send(responseContents);
         });
     });
-
-    webServer.post('/object/*/frame/*/node/*/nodeSize/', function (req, res) {
-        changeNodeSize(req.params[0], req.params[1], req.params[2], req.body, function (statusCode, responseContents) {
-            res.status(statusCode).send(responseContents);
-        });
-    });
-
-    /**
-     * Updates the position and size of a specified node.
-     * @param {string} objectID
-     * @param {string} frameID
-     * @param {string} nodeID
-     * @param {{x: number|undefined, y: number|undefined, scale: number|undefined, matrix: Array.<number>|undefined}} body
-     * @param {function} callback
-     */
-    function changeNodeSize(objectID, frameID, nodeID, body, callback) {
-
-        var updateStatus = 'nothing happened';
-
-        console.log('changing Size for :' + objectID + ' : ' + nodeID);
-
-        getNodeAsync(objectID, frameID, nodeID, function (error, object, frame, node) {
-            if (error) {
-                callback(404, error);
-                return;
-            }
-
-            // check that the numbers are valid numbers..
-            if (typeof body.x === 'number' && typeof body.y === 'number' && typeof body.scale === 'number') {
-                node.x = body.x;
-                node.y = body.y;
-                node.scale = body.scale;
-                updateStatus = 'ok';
-            }
-
-            if (typeof body.matrix === 'object') {
-                node.matrix = body.matrix;
-                updateStatus = 'ok';
-            }
-
-            // if anything updated, write to disk and broadcast updates to editors
-            if (updateStatus === 'ok') {
-                utilities.writeObjectToFile(objects, objectID, objectsPath, globalVariables.saveToDisk);
-                utilities.actionSender({
-                    reloadObject: {object: objectID, frame: frameID, node: nodeID},
-                    lastEditor: body.lastEditor
-                });
-            }
-
-            callback(200, updateStatus);
-        });
-
-    }
 
     // Responds with the set of Spatial Tools that this server is hosting
     webServer.get('/availableFrames/', function (req, res) {
@@ -2143,199 +1731,6 @@ function objectWebServer() {
         return blockList;
     }
 
-    // uploads a new name for a logic node (or could be used for any type of node)
-    // ****************************************************************************************************************
-    webServer.post('/object/:objectID/frame/:frameID/node/:nodeID/rename/', function (req, res) {
-
-        var objectID = req.params.objectID;
-        var frameID = req.params.frameID;
-        var nodeID = req.params.nodeID;
-
-        console.log('received name for', objectID, frameID, nodeID);
-
-        getNodeAsync(objectID, frameID, nodeID, function (error, object, frame, node) {
-            if (error) {
-                res.status(404);
-                res.json(error).end();
-                return;
-            }
-
-            node.name = req.body.nodeName;
-
-            utilities.writeObjectToFile(objects, objectID, objectsPath, globalVariables.saveToDisk);
-
-            res.status(200);
-            res.json({success: true}).end();
-        });
-
-    });
-
-    // uploads a new iconImage for a logic block
-    // ****************************************************************************************************************
-    webServer.post('/object/:objectID/frame/:frameID/node/:nodeID/uploadIconImage', function (req, res) {
-
-        var objectID = req.params.objectID;
-        var frameID = req.params.frameID;
-        var nodeID = req.params.nodeID;
-
-        console.log('received icon image for', objectID, frameID, nodeID);
-
-        getNodeAsync(objectID, frameID, nodeID, function (error, object, frame, node) {
-            if (error) {
-                res.status(404);
-                res.json(error).end();
-                return;
-            }
-
-            var iconDir = objectsPath + '/' + object.name + '/' + identityFolderName + '/logicNodeIcons';
-            if (!fs.existsSync(iconDir)) {
-                fs.mkdirSync(iconDir);
-            }
-
-            var form = new formidable.IncomingForm({
-                uploadDir: iconDir,
-                keepExtensions: true,
-                accept: 'image/jpeg'
-            });
-
-            console.log('created form');
-
-            form.on('error', function (err) {
-                res.status(500);
-                res.send(err);
-                throw err;
-            });
-
-            var rawFilepath = form.uploadDir + '/' + nodeID + '_fullSize.jpg';
-
-            if (fs.existsSync(rawFilepath)) {
-                console.log('deleted old raw file');
-                fs.unlinkSync(rawFilepath);
-            }
-
-            form.on('fileBegin', function (name, file) {
-                console.log('fileBegin loading', name, file);
-                file.path = rawFilepath;
-            });
-
-            console.log('about to parse');
-
-            form.parse(req, function (err, fields) {
-
-                console.log('successfully created icon image', err, fields);
-
-                var resizedFilepath = form.uploadDir + '/' + nodeID + '.jpg';
-                console.log('attempting to write file to ' + resizedFilepath);
-
-                if (fs.existsSync(resizedFilepath)) {
-                    console.log('deleted old resized file');
-                    fs.unlinkSync(resizedFilepath);
-                }
-
-                // copied fullsize file into resized image file as backup, in case resize operation fails
-                fs.copyFileSync(rawFilepath, resizedFilepath);
-
-                if (Jimp) {
-                    Jimp.read(rawFilepath).then(image => {
-                        return image.resize(200, 200).write(resizedFilepath);
-                    }).then(() => {
-                        console.log('done resizing');
-
-                        if (node) {
-                            node.iconImage = 'custom'; //'http://' + object.ip + ':' + serverPort + '/logicNodeIcon/' + object.name + '/' + nodeID + '.jpg';
-                            utilities.writeObjectToFile(objects, objectID, objectsPath, globalVariables.saveToDisk);
-                            utilities.actionSender({
-                                loadLogicIcon: {
-                                    object: objectID,
-                                    frame: frameID,
-                                    node: nodeID,
-                                    ip: object.ip,
-                                    iconImage: node.iconImage
-                                }
-                            }); // TODO: decide whether to send filepath directly or just tell it to reload the logic node from the server... sending directly is faster, fewer side effects
-                        }
-
-                        res.status(200);
-                        res.json({success: true}).end();
-                    }).catch(err => {
-                        console.log('error resizing', err);
-                        res.status(500);
-                        res.send(err);
-                        throw err;
-                    });
-                }
-
-            });
-
-            console.log('parse called');
-
-        });
-
-
-    });
-
-    /**
-     * Normal Links
-     **/
-
-    // delete a link. *1 is the object *2 is the link id
-    // ****************************************************************************************************************
-    webServer.delete('/object/*/link/*/lastEditor/*/', function (req, res) {
-        res.send(deleteLink(req.params[0], req.params[0], req.params[1], req.params[2]));
-    });
-
-    webServer.delete('/object/*/frame/*/link/*/editor/*/deleteLink', function (req, res) {
-        res.send(deleteLink(req.params[0], req.params[1], req.params[2], req.params[3]));
-    });
-
-    /**
-     * Extracts a nicely structured set of data about the link
-     * @param {Link} fullEntry
-     * @param wasAdded
-     * @return {*}
-     */
-    function getLinkData(fullEntry, wasAdded) {
-        var linkAddedData = null;
-
-        if (fullEntry) {
-            console.log('getLinkData', fullEntry);
-
-            var linkObjectA = fullEntry['objectA'];
-            var linkObjectB = fullEntry['objectB'];
-            var linkFrameA = fullEntry['frameA'];
-            var linkFrameB = fullEntry['frameB'];
-            var linkNodeA = fullEntry['nodeA'];
-            var linkNodeB = fullEntry['nodeB'];
-
-            var objectAName = fullEntry['namesA'][0];
-            var objectBName = fullEntry['namesB'][0];
-            var frameAName = fullEntry['namesA'][1];
-            var frameBName = fullEntry['namesB'][1];
-            var nodeAName = fullEntry['namesA'][2]; // TODO: implement a single, safe way to get the object/frame/node (like in the editor) and return null if not found (instead of crashing)
-            var nodeBName = fullEntry['namesB'][2];
-
-            linkAddedData = {
-                added: wasAdded,
-                idObjectA: linkObjectA,
-                idObjectB: linkObjectB,
-                idFrameA: linkFrameA,
-                idFrameB: linkFrameB,
-                idNodeA: linkNodeA,
-                idNodeB: linkNodeB,
-                nameObjectA: objectAName,
-                nameObjectB: objectBName,
-                nameFrameA: frameAName,
-                nameFrameB: frameBName,
-                nameNodeA: nodeAName,
-                nameNodeB: nodeBName
-            };
-
-        } else {
-            console.log('thisObject does not exist');
-        }
-        return linkAddedData;
-    }
-
     function forEachObject(callback) {
         for (var objectKey in objects) {
             if (!objects.hasOwnProperty(objectKey)) continue;
@@ -2356,426 +1751,6 @@ function objectWebServer() {
             callback(frame.nodes[nodeKey], nodeKey);
         }
     }
-
-    /**
-     * Deletes a regular link from the frame it begins from.
-     * Also delete the websocket to this link's destination server IP if it was the last link from this server to that one.
-     * @param {string} objectKey
-     * @param {string} frameKey
-     * @param {string} linkKey
-     * @param {string} editorID
-     */
-    function deleteLink(objectKey, frameKey, linkKey, editorID) {
-
-        var updateStatus = 'nothing happened';
-
-        var foundFrame = getFrame(objectKey, frameKey);
-
-        if (foundFrame) {
-
-            var foundLink = foundFrame.links[linkKey];
-            var destinationIp = knownObjects[foundLink.objectB];
-
-            // notify subscribed interfaces that a new link was DELETED // TODO: make sure this is the right place for this
-            var linkAddedData = getLinkData(foundLink, false);
-            if (linkAddedData) {
-                hardwareAPI.connectCall(linkAddedData.idObjectA, linkAddedData.idFrameA, linkAddedData.idNodeA, linkAddedData);
-                hardwareAPI.connectCall(linkAddedData.idObjectB, linkAddedData.idFrameB, linkAddedData.idNodeB, linkAddedData);
-            }
-
-            delete foundFrame.links[linkKey];
-
-            utilities.writeObjectToFile(objects, objectKey, objectsPath, globalVariables.saveToDisk);
-            utilities.actionSender({reloadLink: {object: objectKey, frame: frameKey}, lastEditor: editorID});
-
-            // iterate over all frames in all objects to see if the destinationIp is still used by another link after this was deleted
-            var checkIfIpIsUsed = false;
-            forEachObject(function (thisObject) {
-                forEachFrameInObject(thisObject, function (thisFrame) {
-                    for (var linkCheckerKey in thisFrame.links) {
-                        if (thisFrame.links[linkCheckerKey].objectB === foundLink.objectB) {
-                            checkIfIpIsUsed = true;
-                        }
-                    }
-                });
-            });
-
-            // if the destinationIp isn't linked to at all anymore, delete the websocket to that server
-            if (foundLink.objectB !== foundLink.objectA && !checkIfIpIsUsed) {
-                delete socketArray[destinationIp];
-            }
-
-            console.log('deleted link: ' + linkKey);
-            updateStatus = 'deleted: ' + linkKey + ' in object: ' + objectKey + ' frame: ' + frameKey;
-        }
-
-        return updateStatus;
-
-    }
-
-    // todo links for programms as well
-    // adding a new link to an object. *1 is the object *2 is the link id
-    // ****************************************************************************************************************
-
-    webServer.post('/object/:objectID/frame/:frameID/link/:linkID/addLink/', function (req, res) {
-        console.log('routed by 2');
-        res.status(200).send(newLink(req.params.objectID, req.params.frameID, req.params.linkID, req.body));
-    });
-
-    webServer.post('/object/*/link/*/', function (req, res) {
-        console.log('routed by 1');
-        res.status(200).send(newLink(req.params[0], req.params[0], req.params[1], req.body));
-    });
-
-    /**
-     * Creates a link on the frame containing the node that the link starts from.
-     * @param {string} objectID
-     * @param {string} frameID
-     * @param {string} linkID
-     * @param {Link} body
-     */
-    function newLink(objectID, frameID, linkID, body) {
-
-        var updateStatus = 'nothing happened';
-
-        var foundFrame = getFrame(objectID, frameID);
-        if (foundFrame) {
-
-            console.log('found frame to add link to');
-
-            // todo the first link in a chain should carry a UUID that propagates through the entire chain each time a change is done to the chain.
-            // todo endless loops should be checked by the time of creation of a new loop and not in the Engine
-            body.loop = (body.objectA === body.objectB &&
-                body.frameA === body.frameB &&
-                body.nodeA === body.nodeB);
-
-            foundFrame.links[linkID] = body;
-
-            if (!body.loop) {
-                console.log('added link: ' + linkID);
-                // write the object state to the permanent storage.
-                utilities.writeObjectToFile(objects, objectID, objectsPath, globalVariables.saveToDisk);
-
-                // check if there are new connections associated with the new link.
-                socketUpdater();
-
-                // notify subscribed interfaces that a new link was DELETED // TODO: make sure this is the right place for this
-                var newLink = foundFrame.links[linkID];
-                var linkAddedData = getLinkData(newLink, true);
-                if (linkAddedData) {
-                    hardwareAPI.connectCall(linkAddedData.idObjectA, linkAddedData.idFrameA, linkAddedData.idNodeA, linkAddedData);
-                    hardwareAPI.connectCall(linkAddedData.idObjectB, linkAddedData.idFrameB, linkAddedData.idNodeB, linkAddedData);
-                }
-
-                // call an action that asks all devices to reload their links, once the links are changed.
-                utilities.actionSender({reloadLink: {object: objectID, frame: frameID}, lastEditor: body.lastEditor});
-
-                updateStatus = 'added';
-            } else {
-                updateStatus = 'found endless Loop';
-            }
-
-        }
-
-        return updateStatus;
-    }
-
-    // Add a new node to an object linked to a frame
-    webServer.post('/object/:objectKey/frame/:frameKey/node/:nodeKey/addNode', function (req, res) {
-        addNodeToFrame(req.params.objectKey, req.params.frameKey, req.params.nodeKey, req, res);
-    });
-
-    webServer.post('/object/:objectKey/node/:nodeKey/', function (req, res) {
-        addNodeToFrame(req.params.objectKey, req.params.objectKey, req.params.nodeKey, req, res);
-    });
-
-    /**
-     * Creates a node on the frame specified frame.
-     * @param {string} objectKey
-     * @param {string} frameKey
-     * @param {string} nodeKey
-     * @param {*} req
-     * @param {*} res
-     * @todo don't pass in req and res, instead callback that triggers status code / err / res
-     */
-    function addNodeToFrame(objectKey, frameKey, nodeKey, req, res) {
-
-        var errorMessage = null;
-
-        var foundObject = getObject(objectKey);
-        if (foundObject) {
-            var foundFrame = getFrame(objectKey, frameKey);
-            if (foundFrame) {
-                foundFrame.nodes[nodeKey] = req.body;
-                utilities.writeObjectToFile(objects, objectKey, objectsPath, globalVariables.saveToDisk);
-                utilities.actionSender({reloadObject: {object: objectKey}, lastEditor: req.body.lastEditor});
-            } else {
-                errorMessage = 'Object ' + objectKey + ' frame ' + frameKey + ' not found';
-            }
-        } else {
-            errorMessage = 'Object ' + objectKey + ' not found';
-        }
-
-        if (errorMessage) {
-            res.status(404).json({failure: true, error: errorMessage}).end();
-        } else {
-            res.status(200).json({success: 'true'}).end();
-        }
-    }
-
-    // adding a new lock to an object. *1 is the object *2 is the datapoint id
-    // ****************************************************************************************************************
-
-    webServer.post('/object/*/nodeLock/*/', function (req, res) {
-        res.send(addNodeLock(req.params[0], req.params[0], req.params[1], req.body));
-    });
-
-    webServer.post('/object/*/frame/*/node/*/addLock/', function (req, res) {
-        res.send(addNodeLock(req.params[0], req.params[1], req.params[2], req.body));
-    });
-
-    /**
-     * Sets a lock password on the specified node.
-     * @param {string} objectKey
-     * @param {string} frameKey
-     * @param {string} nodeKey
-     * @param {{lockPassword: string, lockType: string}} body
-     */
-    function addNodeLock(objectKey, frameKey, nodeKey, body) {
-
-        var updateStatus = 'nothing happened';
-
-        var foundNode = getNode(objectKey, frameKey, nodeKey);
-        if (foundNode) {
-            var previousLockPassword = foundNode.lockPassword;
-            var newLockPassword = body.lockPassword;
-            var previousLockType = foundNode.lockType;
-            var newLockType = body.lockType;
-
-            var isLockActionAllowed = (!previousLockPassword && !!newLockPassword) ||
-                (!!newLockPassword && previousLockPassword === newLockPassword && newLockType !== previousLockType);
-
-            if (isLockActionAllowed) {
-                foundNode.lockPassword = newLockPassword;
-                foundNode.lockType = newLockType;
-
-                utilities.writeObjectToFile(objects, objectKey, objectsPath, globalVariables.saveToDisk);
-                utilities.actionSender({reloadNode: {object: objectKey, frame: frameKey, node: nodeKey}});
-
-                updateStatus = 'added';
-
-            } else {
-                if (previousLockPassword === newLockPassword) {
-                    updateStatus = 'already locked by this user';
-                } else {
-                    updateStatus = 'not authorized to add';
-                }
-            }
-
-        }
-
-        return updateStatus;
-    }
-
-    // delete a lock. *1 is the object *2 is the datapoint id *3 is the encrypted user id
-    // TODO: add robust security to the "password" field
-    // ****************************************************************************************************************
-    webServer.delete('/object/*/frame/*/node/*/password/*/deleteLock', function (req, res) {
-        res.send(deleteNodeLock(req.params[0], req.params[1], req.params[2], req.params[3]));
-    });
-
-    webServer.delete('/object/*/nodeLock/*/password/*/', function (req, res) {
-        res.send(deleteNodeLock(req.params[0], req.params[0], req.params[1], req.params[2]));
-    });
-
-    /**
-     * Removes the lock on the specified node if using the correct password.
-     * @param {string} objectKey
-     * @param {string} frameKey
-     * @param {string} nodeKey
-     * @param {string} password
-     */
-    function deleteNodeLock(objectKey, frameKey, nodeKey, password) {
-
-        var updateStatus = 'nothing happened';
-
-        var foundNode = getNode(objectKey, frameKey, nodeKey);
-        if (foundNode) {
-            if (password === foundNode.lockPassword || (globalVariables.debug && password === 'DEBUG')) { // TODO: remove DEBUG mode
-                foundNode.lockPassword = null;
-                foundNode.lockType = null;
-
-                var object = getObject(objectKey);
-                utilities.writeObjectToFile(objects, object, objectsPath, globalVariables.saveToDisk);
-                utilities.actionSender({reloadNode: {object: objectKey, frame: frameKey, node: nodeKey}});
-
-                updateStatus = 'deleted';
-            } else {
-                updateStatus = 'not authorized to delete';
-            }
-        }
-
-        return updateStatus;
-    }
-
-    // adding a new lock to an object link. *1 is the object *2 is the link id
-    // ****************************************************************************************************************
-    webServer.post('/object/*/linkLock/*/', function (req, res) {
-        res.send(addLinkLock(req.params[0], req.params[0], req.params[1], req.body));
-    });
-
-    webServer.post('/object/*/frame/*/link/*/addLock', function (req, res) {
-        res.send(addLinkLock(req.params[0], req.params[1], req.params[2], req.body));
-    });
-
-    /**
-     * Sets a lock password on the specified link.
-     * @param {string} objectKey
-     * @param {string} frameKey
-     * @param {string} linkKey
-     * @param {{lockPassword: string, lockType: string}} body
-     */
-    function addLinkLock(objectKey, frameKey, linkKey, body) {
-        var updateStatus = 'nothing happened';
-
-        var foundFrame = getFrame(objectKey, frameKey);
-        if (foundFrame) {
-            var foundLink = foundFrame.links[linkKey];
-
-            var previousLockPassword = foundLink.lockPassword;
-            var newLockPassword = body.lockPassword;
-            var previousLockType = foundLink.lockType;
-            var newLockType = body.lockType;
-
-            var isLockActionAllowed = (!previousLockPassword && !!newLockPassword) ||
-                (!!newLockPassword && previousLockPassword === newLockPassword && newLockType !== previousLockType);
-
-            if (isLockActionAllowed) {
-                foundLink.lockPassword = newLockPassword;
-                foundLink.lockType = newLockType;
-
-                var object = getObject(objectKey);
-                utilities.writeObjectToFile(objects, object, objectsPath, globalVariables.saveToDisk);
-                utilities.actionSender({reloadLink: {object: object}});
-
-                updateStatus = 'added';
-
-            } else {
-                if (previousLockPassword === newLockPassword) {
-                    updateStatus = 'already locked by this user';
-                } else {
-                    updateStatus = 'not authorized to add';
-                }
-            }
-        }
-
-        return updateStatus;
-    }
-
-    // delete a lock from a link. *1 is the object *2 is the link id *3 is the encrypted user id
-    // ****************************************************************************************************************
-    webServer.delete('/object/*/linkLock/*/password/*/', function (req, res) {
-        res.send(deleteLinkLock(req.params[0], req.params[0], req.params[1], req.params[2]));
-    });
-
-    webServer.delete('/object/*/frame/*/link/*/password/*/deleteLock', function (req, res) {
-        res.send(deleteLinkLock(req.params[0], req.params[1], req.params[2], req.params[3]));
-    });
-
-    /**
-     * Removes the lock on the specified link if using the correct password.
-     * @param {string} objectKey
-     * @param {string} frameKey
-     * @param {string} linkKey
-     * @param {string} password
-     */
-    function deleteLinkLock(objectKey, frameKey, linkKey, password) {
-
-        var updateStatus = 'nothing happened';
-
-        var foundFrame = getFrame(objectKey, frameKey);
-        if (foundFrame) {
-            var foundLink = foundFrame.links[linkKey];
-            if (password === foundLink.lockPassword || password === 'DEBUG') { // TODO: remove DEBUG mode
-                foundLink.lockPassword = null;
-                foundLink.lockType = null;
-
-                var object = getObject(objectKey);
-                utilities.writeObjectToFile(objects, object, objectsPath, globalVariables.saveToDisk);
-                utilities.actionSender({reloadLink: {object: objectKey}});
-
-                updateStatus = 'deleted';
-            } else {
-                updateStatus = 'not authorized to delete';
-            }
-        }
-
-        return updateStatus;
-    }
-
-    // Update the publicData of a frame when it gets moved from one object to another
-    // ****************************************************************************************************************
-
-    /**
-     * Delete the publicData from every node in the frame
-     */
-    webServer.delete('/object/:objectID/frame/:frameID/publicData', function (req, res) {
-
-        // locate the containing frame in a safe way
-        getFrameAsync(req.params.objectID, req.params.frameID, function (error, object, frame) {
-            if (error) {
-                res.status(404).json(error).end();
-                return;
-            }
-
-            // reset the publicData of each node
-            forEachNodeInFrame(frame, function (node) {
-                node.publicData = {};
-            });
-
-            // save state to object.json
-            utilities.writeObjectToFile(objects, req.params.objectID, objectsPath, globalVariables.saveToDisk);
-
-            res.status(200);
-            res.json({success: true}).end();
-        });
-    });
-
-    /**
-     * Add each publicData from the body to the nodes with the same name specified in the publicData's keys
-     */
-    webServer.post('/object/:objectID/frame/:frameID/publicData', function (req, res) {
-
-        var publicData = req.body.publicData;
-
-        // locate the containing frame in a safe way
-        getFrameAsync(req.params.objectID, req.params.frameID, function (error, object, frame) {
-            if (error) {
-                res.status(404).json(error).end();
-                return;
-            }
-
-            // the keys inside publicData are the names of the node that the data belongs to
-            for (var nodeName in publicData) {
-
-                // find the node with the same name
-                var nodeKey = req.params.frameID + nodeName;
-                if (!frame.nodes.hasOwnProperty(nodeKey)) continue;
-
-                var node = frame.nodes[nodeKey];
-                if (node) {
-                    // and set its public data to the correctly indexed data
-                    node.publicData = publicData[nodeName];
-                }
-            }
-
-            // save state to object.json
-            utilities.writeObjectToFile(objects, req.params.objectID, objectsPath, globalVariables.saveToDisk);
-
-            res.status(200);
-            res.json({success: true}).end();
-        });
-    });
 
     /**
      * Helper function to return the absolute path to the directory that should contain all
@@ -2799,530 +1774,8 @@ function objectWebServer() {
         return videoDir;
     }
 
-    /**
-     * Upload a video file to the object's metadata folder.
-     * The video is stored in a form, which can be parsed and written to the filesystem.
-     * @todo compress video
-     */
-    webServer.post('/object/:id/video/:videoId', function (req, res) {
-        var objectKey = req.params.id;
-        var videoId = req.params.videoId;
-
-        let object = getObject(objectKey);
-        if (!object) {
-            res.status(404).json('Object ' + objectKey + ' not found').end();
-            return;
-        }
-        try {
-            var videoDir = getVideoDir(object.name);
-
-            var form = new formidable.IncomingForm({
-                uploadDir: videoDir,
-                keepExtensions: true,
-                accept: 'video/mp4'
-            });
-
-            console.log('created form for video');
-
-            form.on('error', function (err) {
-                res.status(500).send(err);
-            });
-
-            var rawFilepath = form.uploadDir + '/' + videoId + '.mp4';
-
-            if (fs.existsSync(rawFilepath)) {
-                console.log('deleted old raw file');
-                fs.unlinkSync(rawFilepath);
-            }
-
-            form.on('fileBegin', function (name, file) {
-                file.path = rawFilepath;
-                console.log('fileBegin loading', name, file);
-            });
-
-            form.parse(req, function (err, fields) {
-
-                if (err) {
-                    console.log('error parsing', err);
-                    res.status(500).send(err);
-                    return;
-                }
-
-                console.log('successfully created video file', err, fields);
-
-                res.status(200).json({success: true}).end();
-            });
-        } catch (e) {
-            console.warn('error parsing video upload', e);
-        }
-    });
-
-    /**
-     *
-     *  HANDLE GIT Interface
-     *
-     */
-
-    webServer.post('/object/:id/saveCommit', function (req, res) {
-        if (isMobile) {
-            res.status(500).send('saveCommit unavailable on mobile');
-            return;
-        }
-        var object = getObject(req.params.id);
-        if (object) {
-            git.saveCommit(object, objects, function () {
-                res.status(200);
-                res.json({success: true}).end();
-            });
-        }
-    });
-
-    webServer.post('/object/:id/resetToLastCommit', function (req, res) {
-        if (isMobile) {
-            res.status(500).send('resetToLastCommit unavailable on mobile');
-            return;
-        }
-        var object = getObject(req.params.id);
-        if (object) {
-            git.resetToLastCommit(object, objects, function () {
-                res.status(200);
-                res.json({success: true}).end();
-                hardwareAPI.runResetCallbacks(req.params.id);
-            });
-        }
-    });
-
-    webServer.post('/object/:id/matrix', function (req, res) {
-        let object = getObject(req.params.id);
-        if (!object) {
-            res.status(404).json({
-                failure: true,
-                error: 'Object ' + req.params.id + ' not found'
-            }).end();
-            return;
-        }
-
-        object.matrix = req.body.matrix;
-        console.log('set matrix for ' + req.params.id + ' to ' + object.matrix.toString());
-
-        utilities.writeObjectToFile(objects, req.params.id, objectsPath, globalVariables.saveToDisk);
-
-        res.json({success: true}).end();
-    });
-
-    // Handler of new memory uploads
-    webServer.post('/object/:id/memory', function (req, res) {
-        memoryUpload(req.params.id, /*req.params.id,*/ req, res);
-    });
-
-    // webServer.post('/object/:id/frame/:frame/memory', function (req, res) {
-    //     memoryUpload(req.params.id, req.params.frame, req, res);
-    // });
-
-    /**
-     * Upload an image file to the object's metadata folder.
-     * The image is stored in a form, which can be parsed and written to the filesystem.
-     * @param {string} objectID
-     * @param {express.Request} req
-     * @param {express.Response} res
-     */
-    function memoryUpload(objectID, req, res) {
-
-        if (!objects.hasOwnProperty(objectID)) {
-            res.status(404);
-            res.json({failure: true, error: 'Object ' + objectID + ' not found'}).end();
-            return;
-        }
-
-        var obj = getObject(objectID);
-
-        if (obj.isHumanPose) {
-            res.status(404).json({failure: true, error: 'Object ' + objectID + ' has no directory'}).end();
-            return;
-        }
-
-        var memoryDir = objectsPath + '/' + obj.name + '/' + identityFolderName + '/memory/';
-        if (!fs.existsSync(memoryDir)) {
-            fs.mkdirSync(memoryDir);
-        }
-
-        var form = new formidable.IncomingForm({
-            uploadDir: memoryDir,
-            keepExtensions: true,
-            accept: 'image/jpeg'
-        });
-
-        form.on('error', function (err) {
-            res.status(500);
-            res.send(err);
-            throw err;
-        });
-
-        form.on('fileBegin', function (name, file) {
-            if (name === 'memoryThumbnailImage') {
-                file.path = form.uploadDir + '/memoryThumbnail.jpg';
-            } else {
-                file.path = form.uploadDir + '/memory.jpg';
-            }
-        });
-
-        form.parse(req, function (err, fields) {
-            if (obj) {
-                obj.memory = JSON.parse(fields.memoryInfo);
-                obj.memoryCameraMatrix = JSON.parse(fields.memoryCameraInfo);
-                obj.memoryProjectionMatrix = JSON.parse(fields.memoryProjectionInfo);
-
-                utilities.writeObjectToFile(objects, objectID, objectsPath, globalVariables.saveToDisk);
-                utilities.actionSender({loadMemory: {object: objectID, ip: obj.ip}});
-            }
-
-            console.log('successfully created memory');
-
-            res.status(200);
-            res.json({success: true}).end();
-        });
-    }
-
-    // Create a frame for an object. Assigns it a new UUID.
-    webServer.post('/object/*/frames/', function (req, res) {
-        var frameId = 'frame' + utilities.uuidTime();
-        addFrameToObject(req.params[0], frameId, req.body, res);
-    });
-
-    // Create a frame for an object. Uses its existing UUID
-    webServer.post('/object/*/addFrame/', function (req, res) {
-        var frame = req.body;
-        addFrameToObject(req.params[0], frame.uuid, frame, res);
-    });
-
-    /**
-     * Adds a provided frame to the specified object
-     * @param {string} objectKey
-     * @param {string} frameKey
-     * @param {Frame} frame
-     * @param {*} res
-     */
-    function addFrameToObject(objectKey, frameKey, frame, res) {
-
-        getObjectAsync(objectKey, function (error, object) {
-
-            if (error) {
-                res.status(404).json(error).end();
-                return;
-            }
-
-            if (!frame.src) {
-                res.status(500).json({failure: true, error: 'frame must have src'}).end();
-                return;
-            }
-
-            if (!object.frames) {
-                object.frames = {};
-            }
-
-            utilities.createFrameFolder(object.name, frame.name, __dirname, objectsPath, globalVariables.debug, frame.location);
-
-            var newFrame = new Frame();
-            newFrame.objectId = frame.objectId;
-            newFrame.uuid = frameKey;
-            newFrame.name = frame.name;
-            newFrame.visualization = frame.visualization;
-            newFrame.ar = frame.ar;
-            newFrame.screen = frame.screen;
-            newFrame.visible = frame.visible;
-            newFrame.visibleText = frame.visibleText;
-            newFrame.visibleEditing = frame.visibleEditing;
-            newFrame.developer = frame.developer;
-            newFrame.links = frame.links;
-            newFrame.nodes = frame.nodes;
-            newFrame.location = frame.location;
-            newFrame.src = frame.src;
-            newFrame.width = frame.width;
-            newFrame.height = frame.height;
-
-            // give default values for this node type to each node's public data, if not already assigned
-            for (let key in newFrame.nodes) {
-                if ((!frame.publicData || Object.keys(frame.publicData).length <= 0) && (!newFrame.nodes[key].publicData || Object.keys(newFrame.nodes[key].publicData).length <= 0)) {
-                    newFrame.nodes[key].publicData = JSON.parse(JSON.stringify(nodeTypeModules[newFrame.nodes[key].type].properties.publicData));
-                }
-            }
-
-            object.frames[frameKey] = newFrame;
-
-            utilities.writeObjectToFile(objects, objectKey, objectsPath, globalVariables.saveToDisk);
-            utilities.actionSender({reloadObject: {object: objectKey}, lastEditor: frame.lastEditor});
-
-            // notifies any open screens that a new frame was added
-            hardwareAPI.runFrameAddedCallbacks(objectKey, newFrame);
-
-            res.json({success: true, frameId: frameKey}).end();
-
-        });
-    }
-
-    /**
-     * Creates a copy of the frame (happens when you pull an instance from a staticCopy frame)
-     */
-    webServer.post('/object/:objectID/frames/:frameID/copyFrame/', function (req, res) {
-        var objectID = req.params.objectID;
-        var frameID = req.params.frameID;
-        console.log('making a copy of frame', frameID);
-
-        getFrameAsync(objectID, frameID, function (error, object, frame) {
-            if (error) {
-                res.status(404).json(error).end();
-                return;
-            }
-
-            if (frame.location !== 'global') {
-                return;
-            }
-
-            // don't need to create a folder because we already ensured it is a global frame
-            // (otherwise we would need... utilities.createFrameFolder(object.name, frame.name, ... )
-
-            var newFrame = new Frame();
-            newFrame.objectId = frame.objectId;
-            newFrame.name = frame.src + utilities.uuidTime();
-            var newFrameKey = objectID + newFrame.name;
-            newFrame.uuid = newFrameKey;
-            newFrame.visualization = frame.visualization;
-            // deep clone ar by value, not reference, otherwise posting new position for one might affect the other
-            newFrame.ar = {
-                x: frame.ar.x,
-                y: frame.ar.y,
-                scale: frame.ar.scale,
-                matrix: frame.ar.matrix
-            };
-            // deep clone screen by value, not reference
-            newFrame.screen = {
-                x: frame.screen.x,
-                y: frame.screen.y,
-                scale: frame.screen.scale
-            };
-            newFrame.visible = frame.visible;
-            newFrame.visibleText = frame.visibleText;
-            newFrame.visibleEditing = frame.visibleEditing;
-            newFrame.developer = frame.developer;
-            newFrame.links = frame.links;
-
-            // perform a deep clone of the nodes so it copies by value, not reference
-            newFrame.nodes = {}; // adjust node keys, etc, for copy
-            for (var oldNodeKey in frame.nodes) {
-                if (!frame.nodes.hasOwnProperty(oldNodeKey)) continue;
-                var newNode = new Node();
-                var oldNode = frame.nodes[oldNodeKey];
-                for (var propertyKey in oldNode) {
-                    if (!oldNode.hasOwnProperty(propertyKey)) continue;
-                    newNode[propertyKey] = oldNode[propertyKey];
-                }
-                newNode.frameId = newFrameKey;
-                var newNodeKey = newNode.frameId + newNode.name;
-                newNode.uuid = newNodeKey;
-                newFrame.nodes[newNodeKey] = newNode;
-            }
-
-            newFrame.location = frame.location;
-            newFrame.src = frame.src;
-            newFrame.width = frame.width;
-            newFrame.height = frame.height;
-            object.frames[newFrameKey] = newFrame;
-
-            utilities.writeObjectToFile(objects, objectID, objectsPath, globalVariables.saveToDisk);
-
-            // TODO: by not sending action sender, we assume this is a screen frame -- is that an ok assumption?
-            // utilities.actionSender({reloadObject: {object: objectID}, lastEditor: frame.lastEditor});
-            utilities.actionSender({
-                reloadFrame: {object: objectID, frame: newFrameKey},
-                lastEditor: req.body.lastEditor
-            });
-
-            hardwareAPI.runFrameAddedCallbacks(objectID, newFrame); // creates frame in screen hardware interface
-
-            res.status(200).json({success: true, frameId: newFrameKey, frame: newFrame}).end();
-        });
-
-    });
-
-    /**
-     * Update an object's frame
-     */
-    webServer.post('/object/*/frames/*/', function (req, res) {
-        var objectId = req.params[0];
-        var frameId = req.params[1];
-
-        getObjectAsync(objectId, function (error, object) {
-
-            if (error) {
-                res.status(404).json(error).end();
-                return;
-            }
-
-            var frame = req.body;
-
-            if (!frame.src) {
-                res.status(500).json({failure: true, error: 'frame must have src'}).end();
-                return;
-            }
-
-            if (!object.frames) {
-                object.frames = {};
-            }
-
-            if (!object.frames[frameId]) {
-                object.frames[frameId] = new Frame();
-            }
-
-            frame.loaded = false;
-            // Copy over all properties of frame
-            Object.assign(object.frames[frameId], frame);
-
-            utilities.writeObjectToFile(objects, objectId, objectsPath, globalVariables.saveToDisk);
-
-            utilities.actionSender({reloadObject: {object: objectId}, lastEditor: req.body.lastEditor});
-
-            res.json({success: true}).end();
-
-        });
-
-    });
-
-    /**
-     * Delete a frame from an object
-     */
-    webServer.delete('/object/:objectId/frames/:frameId/', function (req, res) {
-
-        var objectId = req.params.objectId;
-        var frameId = req.params.frameId;
-
-        console.log('delete frame from server', objectId, frameId);
-
-        var object = getObject(objectId);
-        if (!object) {
-            res.status(404).json({failure: true, error: 'object ' + objectId + ' not found'}).end();
-            return;
-        }
-
-        var frame = object.frames[frameId];
-        if (!frame) {
-            res.status(404).json({failure: true, error: 'frame ' + frameId + ' not found'}).end();
-            return;
-        }
-
-        //delete any videos associated with the frame, if necessary
-        // var isPublicDataOnFrame = frame.publicData.hasOwnProperty('data');
-        var publicDataOnAllNodes = Object.keys(frame.nodes).map(function (nodeKey) {
-            return frame.nodes[nodeKey].publicData;
-        });
-        var videoPaths = publicDataOnAllNodes.filter(function (publicData) {
-            if (publicData.hasOwnProperty('data') && typeof publicData.data === 'string') {
-                if (publicData.data.indexOf('http') > -1 && publicData.data.indexOf('.mp4') > -1) {
-                    return true;
-                }
-            }
-            return false;
-        }).map(function (publicData) {
-            return publicData.data;
-        });
-        console.log('frame being deleted contains these video paths: ', videoPaths);
-        videoPaths.forEach(function (videoPath) {
-            // convert videoPath into path on local filesystem // TODO: make this independent on OS path-extensions
-            var urlArray = videoPath.split('/');
-
-            var objectName = urlArray[4];
-            var videoDir = getVideoDir(objectName);
-            var videoFilePath = path.join(videoDir, urlArray[6]);
-
-            if (fs.existsSync(videoFilePath)) {
-                fs.unlinkSync(videoFilePath);
-            }
-        });
-
-        var objectName = object.name;
-        var frameName = object.frames[frameId].name;
-
-        delete object.frames[frameId];
-
-        // remove the frame directory from the object
-        utilities.deleteFrameFolder(objectName, frameName, objectsPath);
-
-        // Delete frame's nodes // TODO: I don't think this is updated for the current object/frame/node hierarchy
-        var deletedNodes = {};
-        for (var nodeId in object.nodes) {
-            var node = object.nodes[nodeId];
-            if (node.frame === frameId) {
-                deletedNodes[nodeId] = true;
-                delete object.nodes[nodeId];
-            }
-        }
-
-        // Delete links involving frame's nodes
-        forEachObject(function (linkObject, linkObjectId) {
-            var linkObjectHasChanged = false;
-
-            for (var linkId in linkObject.links) { // TODO: this isn't updated for frames either
-                var link = linkObject.links[linkId];
-                if (link.objectA === objectId || link.objectB === objectId) {
-                    if (deletedNodes[link.nodeA] || deletedNodes[link.nodeB]) {
-                        linkObjectHasChanged = true;
-                        delete linkObject.links[linkId];
-                    }
-                }
-            }
-
-            if (linkObjectHasChanged) {
-                utilities.writeObjectToFile(objects, linkObjectId, objectsPath, globalVariables.saveToDisk);
-                utilities.actionSender({reloadObject: {object: linkObjectId}, lastEditor: req.body.lastEditor});
-            }
-        });
-
-        // write changes to object.json
-        utilities.writeObjectToFile(objects, objectId, objectsPath, globalVariables.saveToDisk);
-        utilities.actionSender({reloadObject: {object: objectId}, lastEditor: req.body.lastEditor});
-
-        res.json({success: true}).end();
-    });
-
-    // sets the groupID of the specified frame
-    webServer.post('/object/:objectID/frame/:frameID/group/', function (req, res) {
-
-        var frame = getFrame(req.params.objectID, req.params.frameID);
-        if (frame) {
-            var newGroupID = req.body.group;
-            if (newGroupID !== frame.groupID) {
-                frame.groupID = newGroupID;
-                utilities.writeObjectToFile(objects, req.params.objectID, objectsPath, globalVariables.saveToDisk);
-                utilities.actionSender({
-                    reloadFrame: {object: req.params.objectID, frame: req.params.frameID},
-                    lastEditor: req.body.lastEditor
-                });
-                res.status(200).json({success: true}).end();
-                return;
-            }
-        }
-
-        res.status(404).json({success: false, error: 'Couldn\'t find frame ' + req.params.frameID + ' to set groupID'});
-
-    });
-
-    // changing the size and position of an item. *1 is the object *2 is the datapoint id
-
-    // ****************************************************************************************************************
-
     // TODO: is the developer flag ever not true anymore? is it still useful to have?
     if (globalVariables.developer === true) {
-
-        webServer.post('/object/:objectID/frame/:frameID/node/:nodeID/size/', function (req, res) {
-            changeSize(req.params.objectID, req.params.frameID, req.params.nodeID, req.body, function (statusCode, responseContents) {
-                res.status(statusCode).send({status: responseContents});
-            });
-        });
-
-        webServer.post('/object/:objectID/frame/:frameID/size/', function (req, res) {
-            changeSize(req.params.objectID, req.params.frameID, null, req.body, function (statusCode, responseContents) {
-                res.status(statusCode).send({status: responseContents});
-            });
-        });
 
         // // TODO: ask Valentin what this route was used for?
         // webServer.post('/object/*/size/*', function (req, res) {
@@ -3331,123 +1784,6 @@ function objectWebServer() {
         //     res.send(changeSize(req.params[0], req.params[1], null, req.body));
         // });
 
-        /**
-         * Updates the x, y, scale, and/or matrix for the specified frame or node
-         * @todo this function is a mess, fix it up
-         */
-        function changeSize(objectID, frameID, nodeID, body, callback) { // eslint-disable-line no-inner-declarations
-
-            console.log('changing Size for :' + objectID + ' : ' + frameID + ' : ' + nodeID);
-
-            getFrameOrNode(objectID, frameID, nodeID, function (error, object, frame, node) {
-                if (error) {
-                    callback(404, error);
-                    return;
-                }
-
-                var activeVehicle = node || frame; // use node if it found one, frame otherwise
-
-                // console.log('really changing size for ... ' + activeVehicle.uuid, body);
-
-                // console.log("post 2");
-                var updateStatus = 'nothing happened';
-
-                // the reality editor will overwrite all properties from the new frame except these.
-                // useful to not overwrite AR position when sending pos or scale from screen.
-                var propertiesToIgnore = [];
-
-                // TODO: this is a hack to fix ar/screen synchronization, fix it
-                // for frames, the position data is inside "ar" or "screen"
-                if (activeVehicle.hasOwnProperty('visualization')) {
-                    if (activeVehicle.visualization === 'ar') {
-                        activeVehicle = activeVehicle.ar;
-                        propertiesToIgnore.push('screen');
-                    } else if (activeVehicle.visualization === 'screen') {
-                        if (typeof body.scale === 'number' && typeof body.scaleARFactor === 'number') {
-                            activeVehicle.ar.scale = body.scale / body.scaleARFactor;
-                        }
-                        activeVehicle = activeVehicle.screen;
-                        propertiesToIgnore.push('ar.x', 'ar.y'); // TODO: decoding this is currently hard-coded in the editor, make generalized
-                    }
-                }
-
-                var didUpdate = false;
-
-                // check that the numbers are valid numbers..
-                if (typeof body.x === 'number' && typeof body.y === 'number' && typeof body.scale === 'number') {
-
-                    // if the object is equal the datapoint id, the item is actually the object it self.
-                    activeVehicle.x = body.x;
-                    activeVehicle.y = body.y;
-                    activeVehicle.scale = body.scale;
-
-                    if (typeof body.arX === 'number' && typeof body.arY === 'number') {
-                        frame.ar.x = body.arX;
-                        frame.ar.y = body.arY;
-                    }
-
-                    // console.log(req.body);
-                    // ask the devices to reload the objects
-                    didUpdate = true;
-                }
-
-                if (typeof body.matrix === 'object' && activeVehicle.hasOwnProperty('matrix')) {
-                    activeVehicle.matrix = body.matrix;
-                    didUpdate = true;
-                }
-
-                if (didUpdate) {
-                    utilities.writeObjectToFile(objects, objectID, objectsPath, globalVariables.saveToDisk);
-                    utilities.actionSender({
-                        reloadFrame: {
-                            object: objectID,
-                            frame: frameID,
-                            propertiesToIgnore: propertiesToIgnore,
-                            wasTriggeredFromEditor: body.wasTriggeredFromEditor
-                        }, lastEditor: body.lastEditor
-                    });
-                    updateStatus = 'updated position and/or scale';
-                }
-
-                callback(200, updateStatus);
-            });
-
-        }
-
-        webServer.post('/object/*/frame/*/visualization/', function (req, res) {
-            console.log('change visualization');
-            changeVisualization(req.params[0], req.params[1], req.body, res);
-        });
-
-        /**
-         * Sets the visualization to
-         * @param objectKey
-         * @param frameKey
-         * @param { {visualization: string, oldVisualizationPositionData: {{x: number, y: number, scale: number, matrix: Array.<number>}}|undefined } body
-         * @param res
-         */
-        function changeVisualization(objectKey, frameKey, body, res) { // eslint-disable-line no-inner-declarations
-            var newVisualization = body.visualization;
-            var oldVisualizationPositionData = body.oldVisualizationPositionData;
-
-            var frame = getFrame(objectKey, frameKey);
-            if (frame) {
-
-                // if changing from ar -> screen, optionally provide default values for ar.x, ar.y, so that it'll be there when you switch back
-                // if changing from screen -> ar, sets screen.x, screen.y, etc
-                if (oldVisualizationPositionData) {
-                    var oldVisualization = frame.visualization;
-                    frame[oldVisualization] = oldVisualizationPositionData;
-                }
-                frame.visualization = newVisualization;
-
-                utilities.writeObjectToFile(objects, objectKey, objectsPath, globalVariables.saveToDisk);
-
-                res.status(200).json({success: true}).end();
-            } else {
-                res.status(404).json({failure: true, error: 'frame ' + frameKey + ' not found on ' + objectKey}).end();
-            }
-        }
     }
 
     /**
@@ -3518,6 +1854,7 @@ function objectWebServer() {
     // frontend interface
     // ****************************************************************************************************************
 
+    // TODO: is the developer flag ever not true anymore? is it still useful to have?
     if (globalVariables.developer === true) {
 
         // sends the info page for the object :id
@@ -3616,93 +1953,6 @@ function objectWebServer() {
             }
         });
 
-        webServer.get('/object/*/deactivate/', function (req, res) {
-            var objectID = req.params[0];
-            try {
-                getObject(objectID).deactivated = true;
-                utilities.writeObjectToFile(objects, objectID, objectsPath, globalVariables.saveToDisk);
-                res.status(200).send('ok');
-            } catch (e) {
-                res.status(404).json({success: false, error: 'cannot find object with ID' + objectID}).end();
-            }
-        });
-
-        webServer.get('/object/*/activate/', function (req, res) {
-            var objectID = req.params[0];
-            try {
-                getObject(objectID).deactivated = false;
-                utilities.writeObjectToFile(objects, objectID, objectsPath, globalVariables.saveToDisk);
-                res.status(200).send('ok');
-            } catch (e) {
-                res.status(404).json({success: false, error: 'cannot find object with ID' + objectID}).end();
-            }
-        });
-
-        webServer.get('/object/*/screen/', function (req, res) {
-            var objectID = req.params[0];
-            getObject(objectID).visualization = 'screen';
-            console.log('screen', objectID);
-            utilities.writeObjectToFile(objects, objectID, objectsPath, globalVariables.saveToDisk);
-            res.send('ok');
-        });
-
-        webServer.get('/object/*/ar/', function (req, res) {
-            var objectID = req.params[0];
-            getObject(objectID).visualization = 'ar';
-            console.log('ar', objectID);
-            utilities.writeObjectToFile(objects, objectID, objectsPath, globalVariables.saveToDisk);
-            res.send('ok');
-        });
-
-        webServer.get('/object/*/*/reset/', function (req, res) {
-            var objectID = req.params[0];
-            var frameID = req.params[1];
-            var frame = getFrame(objectID, frameID);
-            frame.ar = {
-                x: 0,
-                y: 0,
-                scale: 1,
-                matrix: []
-            };
-            // position data for the screen visualization mode
-            frame.screen = {
-                x: 0,
-                y: 0,
-                scale: 1,
-                matrix: []
-            };
-            utilities.writeObjectToFile(objects, objectID, objectsPath, globalVariables.saveToDisk);
-            res.send('ok');
-        });
-
-        // request a zip-file with the object stored inside
-        // ****************************************************************************************************************
-        webServer.get('/object/:objectId/zipBackup/', function (req, res) {
-            if (isMobile) {
-                res.status(500).send('zipBackup unavailable on mobile');
-                return;
-            }
-            var objectId = req.params.objectId;
-            console.log('sending zipBackup', objectId);
-
-            if (!fs.existsSync(path.join(objectsPath, objectId))) {
-                res.status(404).send('object directory for ' + objectId + 'does not exist at ' + objectsPath + '/' + objectId);
-                return;
-            }
-
-            res.writeHead(200, {
-                'Content-Type': 'application/zip',
-                'Content-disposition': 'attachment; filename=' + objectId + '.zip'
-            });
-
-            var archiver = require('archiver');
-
-            var zip = archiver('zip');
-            zip.pipe(res);
-            zip.directory(objectsPath + '/' + objectId, objectId + '/');
-            zip.finalize();
-        });
-
         // webFrontend realtime messaging
         webServer.post('/webUI/spatial/locator', function (req, res) {
             console.log({
@@ -3726,53 +1976,6 @@ function objectWebServer() {
             console.log('Stop LOG Recording and save file');
             recorder.stop();
             res.status(200).send('ok');
-        });
-
-
-        webServer.post('/object/:objectKey/generateXml/', function (req, res) {
-            var objectKey = req.params.objectKey;
-            var msgObject = req.body;
-            var objectName = msgObject.name;
-
-            console.log(objectKey, msgObject);
-
-            console.log('support inferred aspect ratio of image targets');
-            console.log('support object targets');
-
-            // var isImageTarget = true;
-            // var targetTypeText = isImageTarget ? 'ImageTarget' : 'ObjectTarget'; // not sure if this is actually what object target XML looks like
-
-            var documentcreate = '<?xml version="1.0" encoding="UTF-8"?>\n' +
-                '<ARConfig xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">\n' +
-                '   <Tracking>\n' +
-                '   <ImageTarget name="' + objectKey + '" size="' + parseFloat(msgObject.width).toFixed(8) + ' ' + parseFloat(msgObject.height).toFixed(8) + '" />\n' +
-                '   </Tracking>\n' +
-                '   </ARConfig>';
-
-            let targetDir = path.join(objectsPath, objectName, identityFolderName, 'target');
-            if (!fs.existsSync(targetDir)) {
-                fs.mkdirSync(targetDir);
-                console.log('created directory: ' + targetDir);
-            }
-
-            console.log('am I here!');
-            var xmlOutFile = path.join(targetDir, 'target.xml');
-
-            fs.writeFile(xmlOutFile, documentcreate, function (err) {
-                if (err) {
-                    res.status(500).send('error writing new target size to .xml file for ' + objectKey);
-                } else {
-                    res.status(200).send('ok');
-
-                    // TODO: update object.targetSize.width and object.targetSize.height and write to disk (if object exists yet)
-                    var object = getObject(objectKey);
-                    if (object) {
-                        object.targetSize.width = parseFloat(msgObject.width);
-                        object.targetSize.height = parseFloat(msgObject.height);
-                        utilities.writeObjectToFile(objects, objectKey, objectsPath, globalVariables.saveToDisk);
-                    }
-                }
-            });
         });
 
         webServer.get('/hardwareInterface/:interfaceName/settings/', function (req, res) {
@@ -3973,42 +2176,6 @@ function objectWebServer() {
             });
         });
 
-        webServer.get('/object/:objectKey/disableFrameSharing/', function (req, res) {
-            var objectKey = req.params.objectKey;
-
-            setFrameSharingEnabled(objectKey, false, function (success, errorMessage) {
-                if (success) {
-                    res.status(200).send('ok');
-                } else {
-                    res.status(500).send(errorMessage);
-                }
-            });
-        });
-
-        webServer.get('/object/:objectKey/enableFrameSharing/', function (req, res) {
-            var objectKey = req.params.objectKey;
-
-            setFrameSharingEnabled(objectKey, true, function (success, errorMessage) {
-                if (success) {
-                    res.status(200).send('ok');
-                } else {
-                    res.status(500).send(errorMessage);
-                }
-            });
-        });
-
-        /**
-         * Enable sharing of Spatial Tools from this server to objects on other servers
-         * @todo: see github issue #23 - function is currently unimplemented
-         * @param {string} objectKey
-         * @param {boolean} shouldBeEnabled
-         * @param {successCallback} callback - success, error message
-         */
-        function setFrameSharingEnabled(objectKey, shouldBeEnabled, callback) { // eslint-disable-line no-inner-declarations
-            callback(true);
-            console.warn('TODO: implement frame sharing... need to set property and implement all side-effects / consequences');
-        }
-
         // request a zip-file with the frame stored inside. *1 is the frameName
         // ****************************************************************************************************************
         webServer.get('/frame/:frameName/zipBackup/', function (req, res) {
@@ -4043,39 +2210,6 @@ function objectWebServer() {
             zip.pipe(res);
             zip.directory(framePath, frameName + '/');
             zip.finalize();
-        });
-
-        // sends json object for a specific reality object. * is the object name
-        // ths is the most relevant for
-        // ****************************************************************************************************************
-        webServer.get('/object/:objectID/frame/:frameID/node/:nodeID/', function (req, res) {
-            var node = getNode(req.params.objectID, req.params.frameID, req.params.nodeID);
-            res.json(node || {});
-        });
-
-        // sends json object for a specific reality frame. 1st * is the object name, 2nd * is the frame name
-        // ths is the most relevant for
-        // ****************************************************************************************************************
-        webServer.get('/object/*/frame/*/', function (req, res) {
-            var objectID = req.params[0];
-            var frameID = req.params[1];
-
-            var thisFrame = getFrame(objectID, frameID);
-            if (thisFrame) {
-                res.status(200).json(thisFrame);
-                return;
-            }
-
-            res.status(404).json({
-                failure: true,
-                error: 'Object: ' + objectID + ', frame: ' + frameID + ' not found'
-            }).end();
-        });
-
-        // sends json object for a specific reality object
-        webServer.get('/object/:objectId/', function (req, res) {
-            var object = getObject(req.params.objectId);
-            res.json(object);
         });
 
         /**
